+++ astropy/io/fits/hdu/hdulist.py
@@ -7,7 +7,6 @@ import gzip
 import os
 import shutil
 import sys
-import textwrap
 import warnings
 
 import numpy as np
@@ -215,10 +214,6 @@ class HDUList(list, _Verify):
             self._read_all = self._file.mode == 'ostream'
         else:
             self._read_all = False
-
-        # This is used for book-keeping for backwards compatibility of
-        # accessing unread HDUs after the file has been closed.
-        self._total_read = None
 
         if hdus is None:
             hdus = []
@@ -900,37 +895,8 @@ class HDUList(list, _Verify):
             When `True`, close the underlying file object.
         """
 
-        # If the underlying file object is not closed then we can still
-        # technically lazy-load HDUs
-
-        if closed:
-            # Get the actual current number of HDUs that have been loaded
-            total_read = list.__len__(self)
-
-            # Now read in any remaining HDUs
-            while self._read_next_hdu():
-                pass
-
         try:
-            self._close(output_verify=output_verify, verbose=verbose,
-                        closed=closed)
-        finally:
-            # This will cause deprecation warnings on any future attempt
-            # to access HDUs that were not loaded before closing
-            if closed:
-                self._total_read = total_read
-
-    def _close(self, output_verify='exception', verbose=False, closed=True):
-        """
-        Internal implementation of close() that does not allow further
-        HDUs to be loaded.
-
-        The HDU pre-loading behavior of the explicit close() is currently
-        maintained only for backwards-compatibility.
-        """
-
-        try:
-            if (self._file and self._file.mode in ['append', 'update']
+            if (self._file and self._file.mode in ('append', 'update')
                     and not self._file.closed):
                 self.flush(output_verify=output_verify, verbose=verbose)
         finally:
@@ -1057,45 +1023,17 @@ class HDUList(list, _Verify):
 
         return hdulist
 
-    def _try_while_unread_hdus(self, func, index, *args, **kwargs):
+    def _try_while_unread_hdus(self, func, *args, **kwargs):
         """
         Attempt an operation that accesses an HDU by index/name
         that can fail if not all HDUs have been read yet.  Keep
         reading HDUs until the operation succeeds or there are no
         more HDUs to read.
-
-        The first argument must always be the index of some HDU in the
-        HDUList, but additional arguments may be passed on.
         """
-
-        # TODO: This is a temporary hack for reporting the deprecation
-        # warning on the pre-lazy-loading behavior of allowing HDUs to
-        # be accessed after the file was closed.  It can be removed once
-        # the deprecation period has passed.
-        if self._file and self._file.closed and self._total_read is not None:
-            if index > self._total_read - 1:
-                warnings.warn(textwrap.dedent(
-                    """\
-                    Accessing an HDU after an HDUList is closed, where
-                    that HDU was no read while the HDUList was open
-                    is deprecated.  That is, you did something like:
-
-                        >>> hdulist.close()
-                        >>> print(hdulist[2].header)
-
-                    even though hdulist[2] had not been read yet.  Instead
-                    do:
-
-                        >>> print(hdulist[2].header)
-                        >>> hdulist.close()
-
-                    or open the file with lazy_load_hdus=False to read all
-                    the HDUs into memory immediately upon opening the file.
-                    """), AstropyDeprecationWarning)
 
         while True:
             try:
-                return func(index, *args, **kwargs)
+                return func(*args, **kwargs)
             except Exception:
                 if self._read_next_hdu():
                     continue